Factor out some helpers from elicitations tests
